NEXUS-1289: Nexus Indexer: Store information about classified artifacts in the index
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@3038 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-app/src/main/java/org/sonatype/nexus/maven/tasks/DefaultSnapshotRemover.java
@@ -229,7 +229,7 @@ public class DefaultSnapshotRemover
 
             if ( days > 0 )
             {
-                this.dateThreshold = System.currentTimeMillis() - ( (long) days * 86400000L );
+                this.dateThreshold = System.currentTimeMillis() - ( days * 86400000L );
             }
             else
             {
@@ -516,7 +516,7 @@ public class DefaultSnapshotRemover
                                 null,
                                 null,
                                 null,
-                                false,
+                                false,  // snapshot
                                 false,
                                 null,
                                 false,
